# lib/api-zod/src/generated/types/alunoUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sistema Classe A - API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AlunoUpdateStatus } from './alunoUpdateStatus';

export interface AlunoUpdate {
  /** @minLength 1 */
  nome?: string;
  dataNascimento?: Date;
  /** @minLength 1 */
  nomeResponsavel?: string;
  /** @minLength 1 */
  telefoneResponsavel?: string;
  turmaId?: number;
  status?: AlunoUpdateStatus;
}