# lib/api-zod/src/generated/types/gerarEmMassaResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sistema Classe A - API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GerarEmMassaResult {
  criadas: number;
  ignoradas: number;
  mensagem?: string;
}