# lib/api-zod/src/generated/types/mensalidade.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sistema Classe A - API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MensalidadeStatus } from './mensalidadeStatus';
import type { MensalidadeTipo } from './mensalidadeTipo';

export interface Mensalidade {
  id: number;
  alunoId: number;
  alunoNome: string;
  mes: number;
  ano: number;
  vencimento: Date;
  status: MensalidadeStatus;
  tipo: MensalidadeTipo;
  /** @nullable */
  valor?: number | null;
  /** @nullable */
  dataPagamento?: Date | null;
}