# lib/api-zod/src/generated/types/turmaInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sistema Classe A - API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TurmaInputStatus } from './turmaInputStatus';

export interface TurmaInput {
  /** @minLength 1 */
  nome: string;
  /** @minLength 1 */
  professor: string;
  /** @minLength 1 */
  diasSemana: string;
  /** @minLength 1 */
  horario: string;
  status: TurmaInputStatus;
}